Add files via upload
I got authentication set up but the character minimum is not 8. I set up authentication when I created the project. 
I would need to set the minimum length to 8 in the ApplicationDbContext but didn't have time to figure out where.

I started setting up the relationships between the models from the Model classes. I was not able to set up the many-to-many relationship between the Musicians and Instruments. 
The Musician model has an Icollection<Instrument> but adding an ICollection<Musician> to Instrument caused an error in the DB Migration step. 

I didn't have time to start updating the views to include the database relationships, ie adding Songs to an Album. 

I believe there should be an Address Model since multiple musician could use the same Address but got stuck on how to show that in the View so I added Address and Phone Number to the Muician.

I didn't add an EnsureCreated call to the databases in Program.cs . I don't know if the databases are actually updated into github when I pushed all the code.
The Migrations are updated and I ran Update-Database for each ApplicatonDbContext and NotownRecordsContext.